+++ src/types.rs
@@ -1,16 +1,9 @@
 //! generic types for GPX
 
+use crate::parser::time::Time;
 use geo_types::{Geometry, LineString, MultiLineString, Point, Rect};
-
-#[cfg(feature = "chrono")]
-use chrono::{DateTime, Utc};
 #[cfg(feature = "use-serde")]
 use serde::{Deserialize, Serialize};
-#[cfg(feature = "time")]
-use time::OffsetDateTime as Time;
-
-#[cfg(not(feature = "time"))]
-type Time = DateTime<Utc>;
 
 /// Allowable GPX versions. Currently, only GPX 1.0 and GPX 1.1 are accepted.
 #[derive(Clone, Copy, Debug, PartialEq)]